telegram_woocommerce_integration/requirements.txt
python-telegram-bot
requests
python-dotenv
woocommerce
